backend/pyproject.toml
[tool.bandit]
exclude_dirs = ["tests", ".tox"]

[tool.black]
line-length = 79

[tool.pydocstyle]
convention = "google"
match = ".*\\.py"
match-dir = "(?!migrations).*"
add-ignore = "D107,D104,D106"

[tool.mypy]
plugins = "mypy_django_plugin.main"

[tool.django-stubs]
django_settings_module = "config.settings"

[tool.poetry]
name = "nutrition"
version = "0.1.0"
description = ""
authors = ["Luis Muñoz <luis.munoz.villarreal@gmail.com>"]
packages = [
    { include = "apps" },
    { include = "config" },
    { include = "manage.py" },
]

[tool.poetry.dependencies]
django-admin-sortable2 = "^2.2.1"
django-allow-cidr = "^0.7.1"
django-compressor = "^4.5"
django-cors-headers = "^4.4.0"
django-dbbackup = "^4.1.0"
django-dotenv = "^1.4.2"
django-environ = "^0.11.2"
django-extensions = "^3.2.3"
django-nested-inline = "^0.4.6"
django-storages = {extras = ["google"], version = "^1.14.3"}
django = "^5.0.6"
gunicorn = "^22.0.0"
pint = "^0.23"
psycopg2 = "^2.9.9"
python = "^3.11"
supervisor = "^4.2.5"

[tool.poetry.group.dev.dependencies]
bandit = {extras = ["toml"], version = "^1.7.9"}
black = "^24.4.2"
django-stubs = "^5.0.2"
flake8 = "^7.1.0"
flake8-isort = "^6.1.1"
ipython = "^8.26.0"
ipdb = "^0.13.13"
mypy = "^1.10.1"
pydocstyle = "^6.3.0"
pylint-django = "^2.5.5"
pylint = "^3.2.5"
pytest-cov = "^5.0.0"
pytest-django = "^4.8.0"
pytest-dotenv = "^0.5.2"
pytest-factoryboy = "^2.7.0"
pytest-mock = "^3.14.0"
pytest-xdist = "^3.6.1"
pytest = "^8.2.2"
tox = "^4.16.0"

[tool.poetry.scripts]
nutrition = "manage:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
